chore: refactor monorepo postprocessing utilities
