+++ library_generation/utils/file_render.py
@@ -1,0 +1,24 @@
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     https://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+import os
+from jinja2 import Environment, FileSystemLoader
+
+script_dir = os.path.dirname(os.path.realpath(__file__))
+jinja_env = Environment(loader=FileSystemLoader(f"{script_dir}/../templates"))
+
+
+def render(template_name: str, output_name: str, **kwargs):
+    template = jinja_env.get_template(template_name)
+    t = template.stream(kwargs)
+    directory = os.path.dirname(output_name)
+    if not os.path.isdir(directory):
+        os.makedirs(directory)
+    t.dump(str(output_name))
