+++ library_generation/utils/monorepo_postprocessor.py
@@ -1,0 +1,29 @@
+# Copyright 2024 Google LLC
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     https://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+from library_generation.utils.pom_generator import generate_gapic_bom
+from library_generation.utils.pom_generator import generate_root_pom
+
+
+def monorepo_postprocessing(
+    repository_path: str,
+    versions_file: str,
+) -> None:
+    """
+    Perform repository level post-processing
+    :param repository_path: the path of the repository
+    :param versions_file: the versions_txt contains version of modules
+    :return: None
+    """
+    generate_root_pom(repository_path=repository_path)
+    generate_gapic_bom(repository_path=repository_path, versions_file=versions_file)
